Module Katbv_lib.Semantics

module Env : sig ... end

val of_str_lst : (Base.string * Base.string) Base.list -> Env.t

of_str_lst lst is env where x maps to Bistring.of_binary v in env if (x,v) is in lst

type env_set = Base.Set.M(Env).t

val to_str_lst : env_set -> (Base.string * Base.string) Base.list Base.list

to_str_lst set is the list of elements of the form (x1,s1);...;(xn,sn) for some (x1,b1);...;(xn,bn) in set such that si=Bitstring.to_string bi

val eval : env:Env.t -> Ast.exp -> env_set

eval ~env exp is the set of mappings from field values to bit strings produced by exp evaluated on env. Note that fields unspecified by env are assumed to be 0.